requirements.txt
streamlit==1.46.0
huggingface-hub==0.27.0
requests==2.31.0
sentence-transformers==2.3.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
langchain==0.1.7
watchdog==3.0.0
python-dotenv==1.0.0
Pillow==10.3.0
transformers==4.43.1
PyMuPDF==1.24.2
streamlit-quill==0.0.7
markdownify==0.11.6 